refactor: code splitting resolvers - auth, events, booking, helper
