+++ graphql/resolvers/index.js
@@ -1,148 +1,11 @@
-const bcrypt = require('bcryptjs');
-const Event = require('../../models/event');
-const User = require('../../models/user');
-const Booking = require('../../models/booking');
-const { dateToString } = require('../../helpers/date');
+const authResolver = require("./auth");
+const eventsResolver = require("./events");
+const bookingResolver = require("./booking");
 
-const transformEvent = event => {
-    return {
-        ...event._doc,
-        date: dateToString(event._doc.date),
-        creator: user.bind(this, event._doc.creator)
-    };
+const rootResolver = {
+    ...authResolver,
+    ...eventsResolver,
+    ...bookingResolver
 }
 
-const transformBooking = booking => {
-    return {
-        ...booking._doc,
-        user: user.bind(this, booking._doc.user),
-        event: singleEvent.bind(this, booking._doc.event),
-        createdAt: dateToString(booking._doc.createdAt),
-        updatedAt: dateToString(booking._doc.updatedAt)
-    }
-}
-
-
-// Below functions are basically for drilling i.e. getting collection details based on their id
-const events = async eventIds => {
-    try {
-        const events = await Event.find({ _id: { $in: eventIds } });
-        return events.map(event => {
-            return transformEvent(event);
-        });
-    } catch (err) {
-        throw err;
-    }
-}
-
-const singleEvent = async eventId => {
-    try {
-        const event = await Event.findById(eventId);
-        return transformEvent(event);
-    } catch (err) {
-        throw err;
-    }
-}
-
-const user = async userId => {
-    try {
-        const user = await User.findById(userId);
-        return {
-            ...user._doc,
-            createdEvents: events.bind(this, user._doc.createdEvents)
-        };
-    } catch (err) {
-        throw err;
-    }
-}
-
-module.exports = {
-    events: async () => {
-        try {
-            const events = await Event.find();
-            return events.map(event => {
-                return transformEvent(event);
-            });
-        } catch (err) {
-            throw err;
-        }
-    },
-    bookings: async () => {
-        try {
-            const bookings = await Booking.find();
-
-            return bookings.map(booking => {
-                return transformBooking(booking);
-            });
-        } catch (err) {
-            throw err;
-        }
-    },
-    createEvent: async (args) => {
-        const event = new Event({
-            title: args.eventInput.title,
-            description: args.eventInput.description,
-            price: +args.eventInput.price,
-            date: new Date(args.eventInput.date),
-            creator: '65f9c73e7a2cba2618dafae8'
-        });
-        let createEvent;
-        try {
-            const result = await event.save();
-            createEvent = transformEvent(result);
-            const creator = await User.findById('65f9c73e7a2cba2618dafae8');
-            if (!creator) {
-                throw new Error("User not found!");
-            }
-
-            creator.createdEvents.push(event);
-            await creator.save();
-            return createEvent;
-        } catch (err) {
-            console.log(err);
-        }
-    },
-    createUser: async (args) => {
-        try {
-            const existingUser = await User.findOne({ email: args.userInput.email });
-            if (existingUser) {
-                throw new Error(`User exists already!`);
-            }
-            const hashedPassword = await bcrypt.hash(args.userInput.password, 12);
-            const user = new User({
-                email: args.userInput.email,
-                password: hashedPassword
-            });
-            const result = await user.save();
-            return { ...result._doc, id: result.id, password: null };
-        } catch (err) {
-            throw err;
-        }
-
-
-    },
-    bookEvent: async (args) => {
-        try {
-            const fetchedEvent = await Event.findOne({ _id: args.eventId });
-            const booking = new Booking({
-                user: '65f9c73e7a2cba2618dafae8',
-                event: fetchedEvent
-            });
-            const result = await booking.save();
-
-            return transformBooking(result);
-        } catch (err) {
-            throw err;
-        }
-    },
-    cancelBooking: async (args) => {
-        try {
-            const booking = await Booking.findById(args.bookingId).populate('event');
-            const event = transformEvent(booking.event);
-            await Booking.deleteOne({ _id: args.bookingId });
-            return event;
-        } catch (err) {
-            throw err;
-        }
-    }
-};
+module.exports = rootResolver;
